Introduce GrantPools for concurrent requests
A grant can only have one active ticket at a time. If you fetch a new ticket old tickets become invalid. So if we want to handle multiple requests concurently we need to reserve the grant for the duration of the request. The class GrantPools, helps with managing exclusive access to grants.